D3D11Mesh: did a static_cast to remove warning. Camera / PerspectiveCamera: added useful unprojection methods. Matrix4f: Added infinitePerspectiveProjection that takes in a fov and aspect ratio. libcgt_vecmath: Removed references to non-existent files.

+++ QDirectX/D3D11/src/D3D11Mesh.cpp
@@ -154,7 +154,7 @@ D3D11Mesh::D3D11Mesh( ID3D11Device* pDevice, std::shared_ptr< OBJData > pOBJData
 
 int D3D11Mesh::capacity() const
 {
-	return m_vertexArray.size();
+	return static_cast< int >( m_vertexArray.size() );
 }
 
 std::vector< Vector2i >& D3D11Mesh::vertexRanges()
